# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CROAS OS — Sistema de Gestão 360 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /permissions/{userId}/{module}:
    put:
      operationId: setPermission
      tags: [permissions]
      summary: Set permission level for a user+module cell
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
        - name: module
          in: path
          required: true
          schema:
            type: string
            enum:
              - dashboard
              - finance
              - students
              - files
              - commercial
              - team
              - calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionInput"
      responses:
        "200":
          description: Updated permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionEntry"

  # ─── ACTIVITY LOG ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PermissionEntry:
      type: object
      required: [id, userId, module, level]
      properties:
        id:
          type: number
        userId:
          type: number
        userName:
          type: ["string", "null"]
        module:
          type: string
          enum:
            - dashboard
            - finance
            - students
            - files
            - commercial
            - team
            - calendar
        level:
          type: string
          enum: [NONE, READ, WRITE]
    PermissionInput:
      type: object
      required: [level]
      properties:
        level:
          type: string
          enum: [NONE, READ, WRITE]

    # ─── ACTIVITY LOG ─────────────────────────────────────────────────────────
...